requirements.txt
fastapi
uvicorn
python-docx
jinja2
